requirements.txt
torch
torchvision
matplotlib
numpy
pillow
gradio
scikit-learn
